backend/silobytes/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
django-filter==2.4.0
Markdown==3.3.3
django-rest-knox==4.1.0
environs==9.3.0
